requirements.txt
numpy
pandas
seaborn
neattext
plotly
scikit-learn
joblib
altair
python-docx
pytz